+++ x/fees/keeper/common_test.go
@@ -53,8 +53,6 @@ func testCodec() *codec.Codec {
 
 	// register the different types
 	cdc.RegisterInterface((*crypto.PubKey)(nil), nil)
-	types.RegisterCodec(cdc)
-
 	cdc.Seal()
 	return cdc
 }