+++ x/fees/types/genesis_test.go
@@ -20,8 +20,8 @@ func TestValidateGenesis(t *testing.T) {
 		},
 		{
 			name: "Genesis with invalid params errors",
-			genesis: types.NewGenesisState(types.NewParams("", []types.MinFee{
-				types.NewMinFee("create_post", sdk.NewDecWithPrec(-1, 2))})),
+			genesis: types.NewGenesisState(types.NewParams([]types.MinFee{
+				types.NewMinFee("", sdk.NewCoins(sdk.NewCoin("stake", sdk.NewInt(1))))})),
 			shouldError: true,
 		},
 	}