+++ x/fees/types/min_fee.go
@@ -8,11 +8,11 @@ import (
 
 // MinFee contains the minimum amount of coins that should be paid as a fee for each message of the specified type sent
 type MinFee struct {
-	MessageType string  `json:"message_type" yaml:"message_type"`
-	Amount      sdk.Dec `json:"amount" yaml:"amount"`
+	MessageType string    `json:"message_type" yaml:"message_type"`
+	Amount      sdk.Coins `json:"amount" yaml:"amount"`
 }
 
-func NewMinFee(messageType string, amount sdk.Dec) MinFee {
+func NewMinFee(messageType string, amount sdk.Coins) MinFee {
 	return MinFee{
 		MessageType: messageType,
 		Amount:      amount,
@@ -27,10 +27,6 @@ func (mf MinFee) String() string {
 func (mf MinFee) Validate() error {
 	if mf.MessageType == "" {
 		return fmt.Errorf("invalid minimum fee message type")
-	}
-
-	if mf.Amount.IsNegative() {
-		return fmt.Errorf("minimum fee amout cannot be negative")
 	}
 
 	return nil