+++ x/fees/types/params_test.go
@@ -9,13 +9,13 @@ import (
 )
 
 func TestDefaultParams(t *testing.T) {
-	params := types.NewParams(types.DefaultFeeDenom, types.DefaultMinFees)
+	params := types.NewParams(types.DefaultMinFees)
 	require.Equal(t, params, types.DefaultParams())
 }
 
 func TestParams_String(t *testing.T) {
 	params := types.DefaultParams()
-	require.Equal(t, "Fee parameters:\nFeeDenom: stake\nMinFees: []", params.String())
+	require.Equal(t, "Fee parameters:\nMinFees: []", params.String())
 }
 
 func TestValidateParams(t *testing.T) {
@@ -26,16 +26,11 @@ func TestValidateParams(t *testing.T) {
 		expErr error
 	}{
 		{
-			name:   "invalid fee denom length param returns error",
-			params: types.NewParams("", types.DefaultMinFees),
-			expErr: fmt.Errorf("invalid fee denom param, it shouldn't be empty"),
-		},
-		{
 			name: "invalid min fees param returns error",
-			params: types.NewParams("udaric", []types.MinFee{
-				types.NewMinFee("desmos/createPost", sdk.NewDecWithPrec(-1, 2))},
+			params: types.NewParams([]types.MinFee{
+				types.NewMinFee("", sdk.NewCoins(sdk.NewCoin("stake", sdk.NewInt(1))))},
 			),
-			expErr: fmt.Errorf("minimum fee amout cannot be negative"),
+			expErr: fmt.Errorf("invalid minimum fee message type"),
 		},
 		{
 			name:   "valid params returns no error",
@@ -52,38 +47,6 @@ func TestValidateParams(t *testing.T) {
 	}
 }
 
-func TestValidateFeeDenomParam(t *testing.T) {
-	tests := []struct {
-		name     string
-		feeDenom interface{}
-		expErr   error
-	}{
-		{
-			name:     "invalid param type returns error",
-			feeDenom: sdk.NewInt(10),
-			expErr:   fmt.Errorf("invalid parameter type: 10"),
-		},
-		{
-			name:     "invalid param returns error",
-			feeDenom: "",
-			expErr:   fmt.Errorf("invalid fee denom param, it shouldn't be empty"),
-		},
-		{
-			name:     "valid param returns no errors",
-			feeDenom: "udaric",
-			expErr:   nil,
-		},
-	}
-
-	for _, test := range tests {
-		test := test
-		t.Run(test.name, func(t *testing.T) {
-			err := types.ValidateFeeDenomParam(test.feeDenom)
-			require.Equal(t, test.expErr, err)
-		})
-	}
-}
-
 func TestValidateMinFeesParam(t *testing.T) {
 	tests := []struct {
 		name        string
@@ -96,14 +59,16 @@ func TestValidateMinFeesParam(t *testing.T) {
 			expErr:      fmt.Errorf("invalid parameter type: param"),
 		},
 		{
-			name:        "invalid param returns error",
-			requiredFee: []types.MinFee{types.NewMinFee("desmos/createPost", sdk.NewDec(-1))},
-			expErr:      fmt.Errorf("minimum fee amout cannot be negative"),
+			name: "invalid param returns error",
+			requiredFee: []types.MinFee{types.NewMinFee("",
+				sdk.NewCoins(sdk.NewCoin("stake", sdk.NewInt(1))))},
+			expErr: fmt.Errorf("invalid minimum fee message type"),
 		},
 		{
-			name:        "valid param returns no errors",
-			requiredFee: []types.MinFee{types.NewMinFee("desmos/createPost", sdk.NewDec(1))},
-			expErr:      nil,
+			name: "valid param returns no errors",
+			requiredFee: []types.MinFee{types.NewMinFee("desmos/createPost",
+				sdk.NewCoins(sdk.NewCoin("stake", sdk.NewInt(10000))))},
+			expErr: nil,
 		},
 	}