+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Version 0.14.0
 ## Changes
 - Implemented blocked users checks (#298)
-- Implemented a minimum transactions' fee (#230) 
+- Implemented the possibility to set a minimum messages fee (#230) 
 
 ## Bug fixes
 - Fixed height not working in REST queries (#299)